backend/requirements.txt
fastapi[standard]==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
redis[hiredis]==5.2.1
httpx==0.28.1
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.20
websockets==14.1
chromadb-client==0.6.3
ruff==0.9.3
pytest==8.3.4
pytest-asyncio==0.25.2
